+++ src/main/java/com/getindata/connectors/http/internal/utils/ConfigUtils.java
@@ -3,7 +3,6 @@ package com.getindata.connectors.http.internal.utils;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
-import java.util.stream.Stream;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -79,34 +78,6 @@ public final class ConfigUtils {
         }
 
         return propertyKey.substring(delimiterLastIndex + 1);
-    }
-
-    /**
-     * Flat map a given Map of header name and header value map to an array containing both header
-     * names and values. For example, header map of
-     * <pre>{@code
-     *     Map.of(
-     *     header1, val1,
-     *     header2, val2
-     *     )
-     * }</pre>
-     * will be converter to an array of:
-     * <pre>{@code
-     *      String[] headers = {"header1", "val1", "header2", "val2"};
-     * }</pre>
-     *
-     * @param headerMap mapping of header names to header values
-     * @return an array containing both header names and values
-     */
-    public static String[] toHeaderAndValueArray(Map<String, String> headerMap) {
-        return headerMap
-            .entrySet()
-            .stream()
-            .flatMap(entry -> {
-                String originalKey = entry.getKey();
-                String newKey = ConfigUtils.extractPropertyLastElement(originalKey);
-                return Stream.of(newKey, entry.getValue());
-            }).toArray(String[]::new);
     }
 
     private static <T> void tryAddToConfigMap(